+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### Added
+- log version of `ackee-node-logger` package being used to `pkgVersion` field
+
+### Changed
+- change default serializer for `process` object
+- do not log undefined and empty objects in default serializers
+
 ## [0.2.5] - 2018-09-10
 
 ### Added